Multiple fixes

+++ pkg/player/vlcserver/run.go
@@ -7,6 +7,7 @@ import (
 	"context"
 	"encoding/json"
 	"fmt"
+	"net"
 	"os"
 	"os/exec"
 	"runtime"
@@ -15,9 +16,14 @@ import (
 	child_process_manager "github.com/AgustinSRG/go-child-process-manager"
 	"github.com/facebookincubator/go-belt/tool/experimental/errmon"
 	"github.com/facebookincubator/go-belt/tool/logger"
+	"github.com/xaionaro-go/observability"
 	"github.com/xaionaro-go/player/pkg/player/types"
 	"github.com/xaionaro-go/player/pkg/player/vlcserver/client"
 	"github.com/xaionaro-go/xpath"
+)
+
+const (
+	debugRunVLCInTheSameProcess = true
 )
 
 type VLC struct {
@@ -26,6 +32,16 @@ type VLC struct {
 }
 
 func Run(
+	ctx context.Context,
+	title string,
+) (*VLC, error) {
+	if debugRunVLCInTheSameProcess {
+		return runInTheSameProcess(ctx, title)
+	}
+	return run(ctx, title)
+}
+
+func run(
 	ctx context.Context,
 	title string,
 ) (*VLC, error) {
@@ -40,6 +56,7 @@ func Run(
 		return nil, fmt.Errorf("unable to initialize an stdout pipe: %w", err)
 	}
 	cmd.Env = append(os.Environ(), EnvKeyIsVLCServer+"=1")
+	cmd.Env = append(cmd.Env, fmt.Sprintf("%s=%s", EnvKeyLoggingLevel, logger.FromCtx(ctx).Level().String()))
 	err = child_process_manager.ConfigureCommand(cmd)
 	errmon.ObserveErrorCtx(ctx, err)
 	err = cmd.Start()
@@ -70,6 +87,28 @@ func Run(
 	}, nil
 }
 
+func runInTheSameProcess(
+	ctx context.Context,
+	title string,
+) (*VLC, error) {
+	addrCh := make(chan net.Addr, 1)
+	errCh := make(chan error, 1)
+	observability.Go(ctx, func() {
+		errCh <- runVLCServer(ctx, func(reportedAddr net.Addr) error {
+			addrCh <- reportedAddr
+			return nil
+		})
+	})
+	select {
+	case addr := <-addrCh:
+		return &VLC{
+			Client: client.New(title, addr.String()),
+		}, nil
+	case err := <-errCh:
+		return nil, err
+	}
+}
+
 func (vlc *VLC) SetupForStreaming(
 	ctx context.Context,
 ) error {
@@ -79,6 +118,12 @@ func (vlc *VLC) SetupForStreaming(
 func (vlc *VLC) ProcessTitle(
 	ctx context.Context,
 ) (string, error) {
+	if vlc == nil {
+		return "", fmt.Errorf("vlc == nil")
+	}
+	if vlc.Client == nil {
+		return "", fmt.Errorf("vlc.Client == nil")
+	}
 	return vlc.Client.ProcessTitle(ctx)
 }
 
@@ -86,42 +131,84 @@ func (vlc *VLC) OpenURL(
 	ctx context.Context,
 	link string,
 ) error {
+	if vlc == nil {
+		return fmt.Errorf("vlc == nil")
+	}
+	if vlc.Client == nil {
+		return fmt.Errorf("vlc.Client == nil")
+	}
 	return vlc.Client.OpenURL(ctx, link)
 }
 
 func (vlc *VLC) GetLink(
 	ctx context.Context,
 ) (string, error) {
+	if vlc == nil {
+		return "", fmt.Errorf("vlc == nil")
+	}
+	if vlc.Client == nil {
+		return "", fmt.Errorf("vlc.Client == nil")
+	}
 	return vlc.Client.GetLink(ctx)
 }
 
 func (vlc *VLC) EndChan(
 	ctx context.Context,
 ) (<-chan struct{}, error) {
+	if vlc == nil {
+		return nil, fmt.Errorf("vlc == nil")
+	}
+	if vlc.Client == nil {
+		return nil, fmt.Errorf("vlc.Client == nil")
+	}
 	return vlc.Client.EndChan(ctx)
 }
 
 func (vlc *VLC) IsEnded(
 	ctx context.Context,
 ) (bool, error) {
+	if vlc == nil {
+		return false, fmt.Errorf("vlc == nil")
+	}
+	if vlc.Client == nil {
+		return false, fmt.Errorf("vlc.Client == nil")
+	}
 	return vlc.Client.IsEnded(ctx)
 }
 
 func (vlc *VLC) GetPosition(
 	ctx context.Context,
 ) (time.Duration, error) {
+	if vlc == nil {
+		return 0, fmt.Errorf("vlc == nil")
+	}
+	if vlc.Client == nil {
+		return 0, fmt.Errorf("vlc.Client == nil")
+	}
 	return vlc.Client.GetPosition(ctx)
 }
 
 func (vlc *VLC) GetLength(
 	ctx context.Context,
 ) (time.Duration, error) {
+	if vlc == nil {
+		return 0, fmt.Errorf("vlc == nil")
+	}
+	if vlc.Client == nil {
+		return 0, fmt.Errorf("vlc.Client == nil")
+	}
 	return vlc.Client.GetLength(ctx)
 }
 
 func (vlc *VLC) GetSpeed(
 	ctx context.Context,
 ) (float64, error) {
+	if vlc == nil {
+		return 0, fmt.Errorf("vlc == nil")
+	}
+	if vlc.Client == nil {
+		return 0, fmt.Errorf("vlc.Client == nil")
+	}
 	return vlc.Client.GetSpeed(ctx)
 }
 
@@ -129,12 +216,24 @@ func (vlc *VLC) SetSpeed(
 	ctx context.Context,
 	speed float64,
 ) error {
+	if vlc == nil {
+		return fmt.Errorf("vlc == nil")
+	}
+	if vlc.Client == nil {
+		return fmt.Errorf("vlc.Client == nil")
+	}
 	return vlc.Client.SetSpeed(ctx, speed)
 }
 
 func (vlc *VLC) GetPause(
 	ctx context.Context,
 ) (bool, error) {
+	if vlc == nil {
+		return false, fmt.Errorf("vlc == nil")
+	}
+	if vlc.Client == nil {
+		return false, fmt.Errorf("vlc.Client == nil")
+	}
 	return vlc.Client.GetPause(ctx)
 }
 
@@ -142,6 +241,12 @@ func (vlc *VLC) SetPause(
 	ctx context.Context,
 	pause bool,
 ) error {
+	if vlc == nil {
+		return fmt.Errorf("vlc == nil")
+	}
+	if vlc.Client == nil {
+		return fmt.Errorf("vlc.Client == nil")
+	}
 	return vlc.Client.SetPause(ctx, pause)
 }
 
@@ -151,24 +256,48 @@ func (vlc *VLC) Seek(
 	isRelative bool,
 	quick bool,
 ) error {
+	if vlc == nil {
+		return fmt.Errorf("vlc == nil")
+	}
+	if vlc.Client == nil {
+		return fmt.Errorf("vlc.Client == nil")
+	}
 	return vlc.Client.Seek(ctx, pos, isRelative, quick)
 }
 
 func (vlc *VLC) GetVideoTracks(
 	ctx context.Context,
 ) (types.VideoTracks, error) {
+	if vlc == nil {
+		return nil, fmt.Errorf("vlc == nil")
+	}
+	if vlc.Client == nil {
+		return nil, fmt.Errorf("vlc.Client == nil")
+	}
 	return vlc.Client.GetVideoTracks(ctx)
 }
 
 func (vlc *VLC) GetAudioTracks(
 	ctx context.Context,
 ) (types.AudioTracks, error) {
+	if vlc == nil {
+		return nil, fmt.Errorf("vlc == nil")
+	}
+	if vlc.Client == nil {
+		return nil, fmt.Errorf("vlc.Client == nil")
+	}
 	return vlc.Client.GetAudioTracks(ctx)
 }
 
 func (vlc *VLC) GetSubtitlesTracks(
 	ctx context.Context,
 ) (types.SubtitlesTracks, error) {
+	if vlc == nil {
+		return nil, fmt.Errorf("vlc == nil")
+	}
+	if vlc.Client == nil {
+		return nil, fmt.Errorf("vlc.Client == nil")
+	}
 	return vlc.Client.GetSubtitlesTracks(ctx)
 }
 
@@ -176,6 +305,12 @@ func (vlc *VLC) SetVideoTrack(
 	ctx context.Context,
 	vid int64,
 ) error {
+	if vlc == nil {
+		return fmt.Errorf("vlc == nil")
+	}
+	if vlc.Client == nil {
+		return fmt.Errorf("vlc.Client == nil")
+	}
 	return vlc.Client.SetVideoTrack(ctx, vid)
 }
 
@@ -183,6 +318,12 @@ func (vlc *VLC) SetAudioTrack(
 	ctx context.Context,
 	aid int64,
 ) error {
+	if vlc == nil {
+		return fmt.Errorf("vlc == nil")
+	}
+	if vlc.Client == nil {
+		return fmt.Errorf("vlc.Client == nil")
+	}
 	return vlc.Client.SetAudioTrack(ctx, aid)
 }
 
@@ -190,18 +331,38 @@ func (vlc *VLC) SetSubtitlesTrack(
 	ctx context.Context,
 	sid int64,
 ) error {
+	if vlc == nil {
+		return fmt.Errorf("vlc == nil")
+	}
+	if vlc.Client == nil {
+		return fmt.Errorf("vlc.Client == nil")
+	}
 	return vlc.Client.SetSubtitlesTrack(ctx, sid)
 }
 
 func (vlc *VLC) Stop(
 	ctx context.Context,
 ) error {
+	if vlc == nil {
+		return fmt.Errorf("vlc == nil")
+	}
+	if vlc.Client == nil {
+		return fmt.Errorf("vlc.Client == nil")
+	}
 	return vlc.Client.Stop(ctx)
 }
 
 func (vlc *VLC) Close(
 	ctx context.Context,
 ) error {
-	defer vlc.Cmd.Process.Kill()
+	if vlc == nil {
+		return fmt.Errorf("vlc == nil")
+	}
+	if vlc.Cmd != nil {
+		defer vlc.Cmd.Process.Kill()
+	}
+	if vlc.Client == nil {
+		return fmt.Errorf("vlc.Client == nil")
+	}
 	return vlc.Client.Close(ctx)
 }
